apps/api/requirements.txt
fastapi
uvicorn
google-genai
opik
python-dotenv
